fix: update blog content and add markdown-renderer component
This commit contains an update to the blog content file, replacing the original content with markdown and test examples. Further additions include the implementation of the markdown-renderer component, used for rendering markdown content. It also includes the modification of some component HTML files and changes to the package and package-lock json files to add new dependencies.