feat: add blog-footer component to blog detail
A new component named 'blog-footer' has been added to 'blog-detail'. The blog-footer component receives a blog as an input that it's used to show more details. The import and use of this new component are reflected both in the HTML and TS files, and the corresponding unit test has also been created.