feat: add GitHub and external links to blog posts
Added GitHub and external link icons to each blog post in the portfolio app. This includes modifying the blog types and JSON content to accommodate two new fields: `githubRepoLink` and `otherSourceLink`. Also, updated the UI of the blog info and footer components for better layout and spacing. Removed the unused `encodeURIComponent` attribute from the `BlogInfoComponent`.