feat: add 'bootstrapImage' icon and improve accessibility
'bootstrapImage' icon was added to the blog info component. Furthermore, the accessibility of image elements was improved by adding alternative text and ARIA labels. A description for the hero image in the welcome component was also included. The 'setDefaultImage' method in the default-image service was updated to set alternative text when an image is not found.