fix: update application routing and add NotFoundComponent
This commit includes the implementation of a new NotFoundComponent to manage cases where the user lands on a non-existing page. The application routing has been updated to direct to this 404 page for such scenarios. Further, this commit includes some minor modifications in the MySqlConnection component to handle invalid blogId and some updates in navigation paths in the navbar component for enhanced consistency.