fix: update icons path and remove unused dependency
Updated the path for favicon icons in the portfolio's manifest.json file and removed the unused 'rxjs' dependency from package.json. The icons' previous paths were incorrect, its resolution will improve site aesthetics. The removal of 'rxjs' will decrease the installation time and the bundle size as it was not used throughout the project.