fix: simplify and update GitHub actions YAML
The GitHub actions workflow has been simplified and updated to use an action specifically designed for deploying Angular apps to GitHub pages. Now, our deployment is easier to manage and troubleshoot, since we're leveraging a community-developed action that takes care of many of the details. The workflow no longer includes the Angular and Node.js setup processes, making it simpler and more straightforward.