chore(workflows): change daily operation from merge to push in GitHub workflows
The Github workflows have been updated to push the changes from the develop branch to the master branch daily, instead of merging them. This operation aims to replace the master branch with the latest develop branch, making the master branch always up-to-date with the recently developed features. It's ensured by using the 'reset' git operation along with a force push.